Move to AWS ECR and correct lambda names
