+++ .github/workflows/deploy-lambda.yml
@@ -61,7 +61,7 @@ jobs:
       fail-fast: false
       matrix:
         # Loop over each lambda function
-        lambda: ["globus-auth", "submit_dataset", "submit_status"]
+        lambda: ["auth", "submit", "status"]
 
     needs: test
     runs-on: ubuntu-latest
@@ -88,12 +88,7 @@ jobs:
 
       - name: Login to Amazon ECR
         id: login-ecr
-        uses: aws-actions/amazon-ecr-login@v1
-#      - name: Login to Docker Hub
-#        uses: docker/login-action@v2.2.0
-#        with:
-#          username: ${{ secrets.DOCKER_USERNAME }}
-#          password: ${{ secrets.DOCKER_PASSWORD }}
+        uses: aws-actions/amazon-ecr-login@v1.7.0
 
       - name: Set up QEMU
         uses: docker/setup-qemu-action@v2.2.0
@@ -106,6 +101,6 @@ jobs:
         with:
           context: ./aws
           push: true
-          tags: bengal1/${{ matrix.lambda }}:dockerize
+          tags: ${{ steps.login-ecr.outputs.registry }}/mdf-lambdas/${{ matrix.lambda }}:dockerize
           file: ./aws/Dockerfile
           build-args: LAMBDA_SCRIPT=${{ matrix.lambda }}
