+++ infra/mdf/container_lambdas.tf
@@ -2,7 +2,7 @@ data "aws_ecr_authorization_token" "token" {}
 
 resource "aws_ecr_repository" "mdf-connect-lambda-repo" {
   for_each = local.functions
-  name                 = "lambda-repo/${each.key}"
+  name                 = "mdf-lambdas/${each.key}"
   image_tag_mutability = "MUTABLE"
   image_scanning_configuration {
      scan_on_push = true
