+++ infra/mdf/githuboidc.tf
@@ -36,3 +36,9 @@ resource "aws_iam_role_policy_attachment" "lambda-full-access-policy-attachment"
     role = "${aws_iam_role.github_role.name}"
     policy_arn = "arn:aws:iam::aws:policy/AWSLambda_FullAccess"
 }
+
+resource "aws_iam_role_policy_attachment" "ecr-full-access-policy-attachment" {
+    role = "${aws_iam_role.github_role.name}"
+    policy_arn = "arn:aws:iam::aws:policy/AmazonEC2ContainerRegistryFullAccess"
+}
+
